+++ src/commands/github-authors.js
@@ -1,6 +1,8 @@
 const vscode = require("vscode");
-const { get } = require("../github/github-api");
-const { saveNewCoAuthors, fetchGitHubAuthors } = require("git-mob-core");
+const {
+  saveNewCoAuthors,
+  searchGitHubAuthors: gmSearchGhAuthors,
+} = require("git-mob-core");
 
 function searchGithubAuthors() {
   return vscode.commands.registerCommand(
@@ -17,27 +19,23 @@ function searchGithubAuthors() {
       });
 
       if (typeof searchText === "undefined") return null;
-
-      const searchUsers = await get("search/users?q=" + searchText);
       let users = [];
-      if (searchUsers.statusCode <= 400) {
-        const logins = searchUsers.data.items.map((item) => item.login);
-        users = await fetchGitHubAuthors(logins, "vs-code-git-mob");
-      } else {
-        vscode.window.showErrorMessage(
-          "Request to GitHub failed: " + searchUsers.data.message
-        );
+      try {
+        users = await gmSearchGhAuthors(searchText, "vs-code-git-mob");
+      } catch (error) {
+        vscode.window.showErrorMessage(error.message);
         return;
       }
 
-      if (searchUsers.data.total_count === 0) {
+      if (users.length === 0) {
         vscode.window.showInformationMessage("No users found!");
         return;
       }
-      const messageUnder30 = `Git Mob: Showing ${searchUsers.data.total_count} GitHub users.`;
-      const messageOver30 = `Git Mob: Can only showing 30 of ${searchUsers.data.total_count} GitHub users.`;
+
+      const messageUnder30 = `Git Mob: Found ${users.length} GitHub users.`;
+      const messageOver30 = `Git Mob: Can only showing 30 GitHub users but more were found.`;
       vscode.window.showInformationMessage(
-        searchUsers.data.total_count > 30 ? messageOver30 : messageUnder30
+        users.length === 30 ? messageOver30 : messageUnder30
       );
 
       const selectedAuthor = await quickPickAuthors(users);