+++ src/ext-config/config.js
@@ -1,10 +1,5 @@
 const { workspace, commands } = require("vscode");
 
-function getGitHubPat() {
-  const gitHubSettings = workspace.getConfiguration("gitMob.gitHub");
-  return gitHubSettings.get("personalAccessToken");
-}
-
 function getConfigSolo() {
   const config = workspace.getConfiguration("gitMob.postCommit");
   return config.get("solo");
@@ -26,7 +21,6 @@ workspace.onDidChangeConfiguration((evt) => {
   }
 });
 
-exports.getGitHubPat = getGitHubPat;
 exports.getConfigSolo = getConfigSolo;
 exports.getSortDirection = getSortDirection;
 exports.moveToCoAuthoring = moveToCoAuthoring;