+++ CHANGELOG.md
@@ -4,9 +4,15 @@ Follows [Semantic Versioning](https://semver.org/).
 
 ## Next version
 
+## 1.20.1
+
 ### Added
 
 - Use `git-mob-core` to search GitHub users replacing what is in the extension. The limitation is core API does not allow a PAT key and won't get authors selected email (will just return GH anonymous email). Less code to maintain across tools.
+
+### Fix
+
+- When no path set for the commit template, default to the global template. [Issue 302](https://github.com/rkotze/git-mob-vs-code/issues/302)
 
 ## 1.20.0